+++ lib/collection_picker.dart
@@ -2,5 +2,7 @@ library collection_picker;
 
 export 'src/listview_picker.dart';
 export 'src/gridview_picker.dart';
+export 'src/sliver_listview_picker.dart';
+export 'src/sliver_gridview_picker.dart';
 export 'src/picker_typedef.dart';
 export 'src/picker_wrapper.dart';