+++ lib/src/list_extension.dart
@@ -1,7 +1,10 @@
+import 'package:flutter/cupertino.dart';
+
 /// List extension of this collection_picker package
 extension ListExt<T> on List<T> {
   /// List extension to safely find the first data in the list.
   /// If the data is not found, it will return null.
+  @protected
   T? firstWhereOrNull(bool Function(T) test) {
     final index = indexWhere(test);
     if (index != -1) {